+++ handlers.js
@@ -1,44 +1,52 @@
 const knownCommands = require('./functions/index');
 const commandPrefix = require('./config').prefix;
 const version = require('./package').version;
-const sites = ['https://www.keyforgegame.com/deck-details/',
-	'https://decksofkeyforge.com/decks/',
-	'https://keyforge-compendium.com/decks/',
-	'https://burgertokens.com/pages/keyforge-deck-analyzer?deck='
-];
+const _ = require('lodash');
 
 
 // Called every time a message comes in:
 const onMessage = (msg, client) => {
 	if (msg.author.bot) return; // Ignore messages from the bot
-	let params = [], commandName, flags = [];
-
-
-	if (sites.some(a => msg.content.startsWith(a))) {
-		const [site] = sites.filter(a => msg.content.startsWith(a));
-		msg.content = `${commandPrefix}d ${msg.content.replace(site, '').slice(0, 36)}`;
-		flags = ['delete', ...flags];
-	}
-	// This isn't a command since it has no prefix:
-	if (!msg.content.includes(commandPrefix) && !msg.content.match(/[\[\]{}]/g)) return;
-
-	flags = [...flags, ...msg.content.split(' ').filter(a => a.startsWith('-')).map(flag => flag.slice(1))];
-	msg.content = msg.content.split(' ').filter(a => !a.startsWith('-')).join(' ');
-
-	if (msg.content.includes(commandPrefix)) {
+	let params = msg.content.split(' '),
+		commandName = params.map(a => a.startsWith(commandPrefix) && a).filter(Boolean).join().slice(1),
+		flags = params.filter(a => a.startsWith('-')).map(flag => flag.slice(1)),
+		types = {
+			brackets: {
+				regex: /\[(.*?)\]/,
+				params: []
+			},
+			d: {
+				regex: /[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}/,
+				params: []
+			},
+			deck: {
+				regex: /\{(.*?)\}/,
+				params: []
+			},
+		};
+	if (commandName) {
 		if (msg.content.startsWith(commandPrefix)) params = msg.content.substr(1).split(' ');
 		else params = msg.content.split(commandPrefix)[1].split(' ');
-		commandName = params[0].toLowerCase();
-		params = params.splice(1);
-	} else if (msg.content.match(/]/g) && msg.content.match(/\[/g)) {
-		let parse = msg.content.split('[');
-		parse.forEach(sub => sub.includes(']') && params.push(sub.split(']')[0]));
-		commandName = 'brackets';
-	} else if (msg.content.match(/}/g) && msg.content.match(/{/g)) {
-		let parse = msg.content.split('{');
-		parse.forEach(sub => sub.includes('}') && params.push(sub.split('}')[0]));
-		commandName = 'deck';
+		params = params.splice(1).filter(a => !a.startsWith('-'));
 	}
+
+	Object.keys(types).forEach(a => {
+		let message = msg.content;
+		do {
+			let param = message.match(types[a].regex);
+			if (param) {
+				types[a].params.push(_.get(param, '1', param[0]));
+				message = message.replace(_.get(param, '0'), '');
+			}
+		} while (message.match(types[a].regex));
+		if (types[a].params.length > 0) {
+			commandName = a;
+			if (params < 1) flags = [...flags, 'delete'];
+			params = types[a].params
+		}
+	});
+
+	if (!commandName) return;
 
 	switch (commandName) {
 		case 'c':